Update to 4.20.0 and more accurate playtime variables

+++ paper/src/main/java/rocks/gravili/notquests/paper/structs/variables/PlayerPlaytimeHoursVariable.java
@@ -24,25 +24,25 @@ import rocks.gravili.notquests.paper.structs.QuestPlayer;
 
 import java.util.List;
 
-public class PlayerPlaytimeHoursVariable extends Variable<Integer> {
+public class PlayerPlaytimeHoursVariable extends Variable<Double> {
     public PlayerPlaytimeHoursVariable(NotQuests main) {
         super(main);
         setCanSetValue(true);
     }
 
     @Override
-    public Integer getValue(QuestPlayer questPlayer, Object... objects) {
+    public Double getValue(QuestPlayer questPlayer, Object... objects) {
         if (questPlayer != null) {
-            return questPlayer.getPlayer().getStatistic(Statistic.PLAY_ONE_MINUTE)/72000;
+            return questPlayer.getPlayer().getStatistic(Statistic.PLAY_ONE_MINUTE)/72000d;
         } else {
             return null;
         }
     }
 
     @Override
-    public boolean setValueInternally(Integer newValue, QuestPlayer questPlayer, Object... objects) {
+    public boolean setValueInternally(Double newValue, QuestPlayer questPlayer, Object... objects) {
         if (questPlayer != null) {
-            questPlayer.getPlayer().setStatistic(Statistic.PLAY_ONE_MINUTE, newValue*72000);
+            questPlayer.getPlayer().setStatistic(Statistic.PLAY_ONE_MINUTE, newValue.intValue()*72000);
             return true;
         } else {
             return false;
