+++ paper/src/main/java/rocks/gravili/notquests/paper/structs/variables/PlayerPlaytimeMinutesVariable.java
@@ -24,25 +24,25 @@ import rocks.gravili.notquests.paper.structs.QuestPlayer;
 
 import java.util.List;
 
-public class PlayerPlaytimeMinutesVariable extends Variable<Integer> {
+public class PlayerPlaytimeMinutesVariable extends Variable<Double> {
     public PlayerPlaytimeMinutesVariable(NotQuests main) {
         super(main);
         setCanSetValue(true);
     }
 
     @Override
-    public Integer getValue(QuestPlayer questPlayer, Object... objects) {
+    public Double getValue(QuestPlayer questPlayer, Object... objects) {
         if (questPlayer != null) {
-            return questPlayer.getPlayer().getStatistic(Statistic.PLAY_ONE_MINUTE)/1200;
+            return questPlayer.getPlayer().getStatistic(Statistic.PLAY_ONE_MINUTE)/1200d;
         } else {
             return null;
         }
     }
 
     @Override
-    public boolean setValueInternally(Integer newValue, QuestPlayer questPlayer, Object... objects) {
+    public boolean setValueInternally(Double newValue, QuestPlayer questPlayer, Object... objects) {
         if (questPlayer != null) {
-            questPlayer.getPlayer().setStatistic(Statistic.PLAY_ONE_MINUTE, newValue*1200);
+            questPlayer.getPlayer().setStatistic(Statistic.PLAY_ONE_MINUTE, newValue.intValue()*1200);
             return true;
         } else {
             return false;
